+++ lib/searchkick.rb
@@ -223,9 +223,6 @@ module Searchkick
       elsif records.respond_to?(:queryable)
         # Mongoid 3+
         records.queryable.for_ids(ids)
-      elsif records.respond_to?(:unscoped) && :id.respond_to?(:in)
-        # Nobrainer
-        records.unscoped.where(:id.in => ids)
       elsif records.respond_to?(:key_column_names)
         records.where(records.key_column_names.first => ids)
       end