+++ lib/searchkick/model.rb
@@ -99,12 +99,6 @@ module Searchkick
         def should_index?
           true
         end unless method_defined?(:should_index?)
-
-        if defined?(Cequel) && self < Cequel::Record && !method_defined?(:destroyed?)
-          def destroyed?
-            transient?
-          end
-        end
       end
     end
   end