+++ test/index_test.rb
@@ -87,7 +87,6 @@ class IndexTest < Minitest::Test
   end
 
   def test_large_value
-    skip if nobrainer?
     large_value = 1000.times.map { "hello" }.join(" ")
     store [{name: "Product A", text: large_value}], Region
     assert_search "product", ["Product A"], {}, Region
@@ -97,7 +96,6 @@ class IndexTest < Minitest::Test
   end
 
   def test_very_large_value
-    skip if nobrainer?
     large_value = 10000.times.map { "hello" }.join(" ")
     store [{name: "Product A", text: large_value}], Region
     assert_search "product", ["Product A"], {}, Region