+++ test/inheritance_test.rb
@@ -1,11 +1,6 @@
 require_relative "test_helper"
 
 class InheritanceTest < Minitest::Test
-  def setup
-    skip if cequel?
-    super
-  end
-
   def test_child_reindex
     store_names ["Max"], Cat
     assert Dog.reindex