+++ test/reindex_test.rb
@@ -37,8 +37,6 @@ class ReindexTest < Minitest::Test
   end
 
   def test_relation_inline
-    skip if nobrainer? || cequel?
-
     store_names ["Product A"]
     store_names ["Product B", "Product C"], reindex: false
     Product.where(name: "Product B").reindex(refresh: true)
@@ -46,8 +44,6 @@ class ReindexTest < Minitest::Test
   end
 
   def test_relation_associations
-    skip if nobrainer? || cequel?
-
     store_names ["Product A"]
     store = Store.create!(name: "Test")
     Product.create!(name: "Product B", store_id: store.id)
@@ -56,8 +52,6 @@ class ReindexTest < Minitest::Test
   end
 
   def test_relation_should_index
-    skip if nobrainer? || cequel?
-
     skip "TODO make pass in Searchkick 5"
 
     store_names ["Product A", "Product B"]